backend/requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
pydantic==2.13.3
google-generativeai==0.3.1
python-dotenv==1.2.2
deepgram-sdk==3.0.0
python-multipart==0.0.27
